add : generate function on Consensus / entry_genesis
