--- new-rust/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "rust"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-tokio = { version = "1.0", features = ["full"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-
-toml = "0.9.2"
-
-anyhow = "1.0"
-thiserror = "1.0"
-base64 = "0.21"
-hex = "0.4"
-chrono = { version = "0.4", features = ["serde"] }
-once_cell = "1.21.3"
-dashmap = "6.1.0"
-rand = "0.9.2"
-blake3 = "1.8.2"
-bincode = "1.3.3"
-blst = "0.3.15"
-rocksdb = "0.24.0"
-bitvec = "1.0.1"
-reqwest = { version = "0.11", features = ["json", "stream", "blocking", "rustls-tls"] }
-bs58 = "0.5.1"
-rayon = "1.11.0"
-indicatif = "0.17"
-zip = "0.6"
-futures-util = "0.3"
-lazy_static = "1.5.0"
-bls12_381 = "0.8.0"
-group = "0.13.0"
