--- new-rust/src/core/consensus/mod.rs
@@ -1,8 +1,0 @@
-pub mod attestation;
-pub mod bls12_aggsig;
-pub mod entry;
-pub mod entry_genesis;
-pub use attestation::*;
-pub use bls12_aggsig::*;
-pub use entry::*;
-pub use entry_genesis::*;
