--- new-rust/src/core/external/bls_rs/mod.rs
@@ -1,4 +1,0 @@
-pub mod bls_rs;
-pub mod errors;
-pub use bls_rs::*;
-pub use errors::*;
