--- new-rust/src/core/external/mod.rs
@@ -1,3 +1,0 @@
-pub mod bls_rs;
-
-pub use bls_rs::*;
