--- new-rust/src/core/mod.rs
@@ -1,17 +1,0 @@
-pub mod api;
-pub mod bic;
-pub mod consensus;
-pub mod external;
-pub mod http;
-pub mod misc;
-pub mod node;
-pub mod trainer;
-
-pub use api::*;
-pub use bic::*;
-pub use consensus::*;
-pub use external::*;
-pub use http::*;
-pub use misc::*;
-pub use node::*;
-pub use trainer::*;
