--- new-rust/src/lib.rs
@@ -1,9 +1,0 @@
-pub mod config;
-pub mod core;
-pub mod html;
-pub mod rel;
-
-pub use config::*;
-pub use core::*;
-pub use html::*;
-pub use rel::*;
