+++ rust/Config.toml
@@ -31,8 +31,8 @@ trustfactor = 0.8                   # override via TRUSTFACTOR env
 
 # # Trainer keys (generated at runtime)
 # trainer_pk_b58 = ""
-trainer_pk = []
-trainer_sk = []
+trainer_pk = [154,214,92,178,161,190,11,25,154,178,44,19,238,209,142,102,244,249,88,32,46,199,91,87,232,228,149,53,59,119,76,209]
+trainer_sk = [247,240,106,38,68,98,2,252,75,224,21,169,43,196,207,96,96,250,232,190,20,73,177,35,52,64,93,88,121,127,234,80,154,214,92,178,161,190,11,25,154,178,44,19,238,209,142,102,244,249,88,32,46,199,91,87,232,228,149,53,59,119,76,209]
 # trainer_pop = ""
 
 # Optional flags
